eighth commit

+++ src/test/java/univ/lecture/CalculatorTest.java
@@ -10,7 +10,7 @@ public class CalculatorTest {
     @Test
     public void testCalculatorAdd () {
     	Application calc = new Application();
-        int output = (int) calc.calculator("1+1");
+        double output = calc.calculator("1+1");
         assertThat(output, is(2));
     }
 }
